Split tests of mixed into sync and non-sync
